requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
gradio-client==1.3.0
boto3==1.35.18
requests
Pillow

